dashboard/requirements.txt
PyQt6>=6.5
pyqtgraph>=0.13
pyopengl>=3.1
bleak>=0.21
numpy>=1.24
pyserial>=3.5